refactor: Update issuer names in OTP cards
- Update issuer name from "Google" to "Steam"
- Update issuer name from "LinkedIn" to "Epic Games"

Refactor the Home component in the app/page.tsx file to update the issuer names in the OTP cards. The issuer name "Google" has been changed to "Steam", and the issuer name "LinkedIn" has been changed to "Epic Games". This change improves the accuracy and relevance of the issuer names displayed in the OTP cards.